refactor!: move request helpers into a trait
